Crate mimalloc

A drop-in global allocator wrapper around the mimalloc allocator. Mimalloc is a general purpose, performance oriented allocator built by Microsoft.

Usage

use mimalloc::MiMalloc;

#[global_allocator]
static GLOBAL: MiMalloc = MiMalloc;

Usage without secure mode

By default this library builds mimalloc in secure mode. This means that heap allocations are encrypted, but this results in a 3% increase in overhead.

To disable secure mode, in Cargo.toml:

[dependencies]
mimalloc = { version = "*", default-features = false }

Structs

MiMalloc

Drop-in mimalloc global allocator.